% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMR-main.R
\name{construct_analysis}
\alias{construct_analysis}
\title{Internal Method for Constructing IMR Analysis}
\usage{
construct_analysis(mc, X, ...)
}
\arguments{
\item{...}{}
}
\value{
a function that has input \code{Y}.
}
\description{
An internal method used in the \code{\link{IMR}} for construction of the
model analysis.
}
\details{
LONG DESCRIPTION
}
