% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMR-main.R
\name{find_data}
\alias{find_data}
\title{Internal Method for Extracting Data for IMR}
\usage{
find_data(m)
}
\arguments{
\item{m}{a model object}
}
\value{
a list with the following three entries
  \item{X} the fitted values of the model. In hazard models this is the
    time points corresponding to observed events.
  \item{Y} a list with the raw and standerdized residuals. In hazard models
    this is the standerdized residuals found when fitting a line through the
    estimated cumulative hazard.
  \item{name} a ("short") string describing the model. This value will be
    used as the model name in the final output.
}
\description{
An internal generic method for extracting fitted response and standerdize
residuals.
}
\details{
LONG DESCRIPTION
}
