% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICP-fitters.R
\name{fit_model}
\alias{fit_model}
\alias{fit_model.default}
\alias{fit_model.lm}
\alias{fit_model.glm}
\alias{fit_model.ph}
\alias{fit_model.ah}
\alias{fit_model.hazard}
\alias{fit_nonparam_model}
\alias{fit_nonparam_model.default}
\alias{fit_nonparam_model.ph}
\alias{fit_nonparam_model.ah}
\title{Internal regression functions}
\usage{
fit_model(method, Y, X, subset, ...)

\method{fit_model}{default}(method, Y, X, subset = NULL, ...)

\method{fit_model}{lm}(method, Y, X, subset = NULL, ...)

\method{fit_model}{glm}(method, Y, X, subset = NULL, ...)

\method{fit_model}{ph}(method, Y, X, subset = NULL, ...)

\method{fit_model}{ah}(method, Y, X, subset = NULL, ...)

\method{fit_model}{hazard}(method, Y, X, subset = NULL, ...)

fit_nonparam_model(method, Y, X, subset, ...)

\method{fit_nonparam_model}{default}(method, Y, X, subset = NULL, ...)

\method{fit_nonparam_model}{ph}(method, Y, X, subset = NULL, id = NULL, ...)

\method{fit_nonparam_model}{ah}(method, Y, X, subset = NULL, id = NULL, ...)
}
\arguments{
\item{method}{a \strong{method object} created by the
\code{\link{method_obj}} function.}

\item{Y}{a vector or \code{\link[survival]{Surv}} object describing the
target variable.}

\item{X}{a matrix, vector or data frame describing the covariates.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{...}{additional arguments to be passed to lower level functions.}

\item{id}{for timevarying covariates the variable must associate each record
with the id of a subject.}
}
\value{
Both the \code{fit_model} and \code{fit_nonparam_model} methods return a list.

\code{fit_model} methods must return the following:
  \item{\code{coefficients} }{ The estimated regression coefficients.}
  \item{\code{deviance} }{ The deviance of the fitted model, i.e. \eqn{-2 log(likelihood)}}
  \item{\code{df} }{ Degrees of freedom in fitted model.}
  \item{\code{scale} }{ Scale.}


\code{fit_nonparam_model} methods return the following:
  \item{\code{cum} }{ Cumulative regression coefficient.}
  \item{\code{cum.var} }{ Cumulative variance of regression effects.}
  \item{\code{sup} }{ Kolmogorov–Smirnov test (only returned if \code{n.sim} not zero).}
  \item{\code{int} }{ Cramér–von Mises test (only returned if \code{n.sim} not zero).}
}
\description{
The \code{fit_model} function is a generic function meant for internal use in
the \code{ICPSurv} package, and as such it is not exported to the namespace.
}
\details{
The \code{fit_model.X} and \code{fit_nonparam_model.X} functions are internal
fitter functions. They are usually all quite simple as their main
functionality apart from fitting the model is ensureing that the output
is compatible with the \code{\link{plausible_predictor_test}}s.

All \code{fit_model} methods must return the following:
\itemize{
  \item{\code{coefficients} }{ The estimated regression coefficients.}
  \item{\code{deviance} }{ The deviance of the fitted model, i.e. \eqn{-2 log(likelihood)}}
  \item{\code{df} }{ Degrees of freedom in fitted model.}
  \item{\code{scale} }{ Scale.}
}

All \code{fit_nonparam_model} methods must return the following:
\itemize{
  \item{\code{cum} }{ Cumulative regression coefficient.}
  \item{\code{cum.var} }{ Cumulative variance of regression effects.}
}
Morover if \code{n.sim} in the \code{method} is non-zero then the
\code{fit_nonparam_model} also returnes the following
\itemize{
  \item{\code{sup} }{ Kolmogorov–Smirnov test.}
  \item{\code{int} }{ Cramér–von Mises test.}
}
}
\seealso{
\code{\link{ICP}}, \code{\link{plausible_predictor_test}} for the
  full wrapper functions.
}
\keyword{internal}
