% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICP-main.R
\name{ICP}
\alias{ICP}
\title{Invariant Causal Prediction}
\usage{
ICP(Y, X, E = NULL, method = structure(list(model = "glm", family =
  "gaussian"), class = "EnvirIrrel"), level = NULL, fullAnalysis = T,
  maxNoVariables = 8, stopIfEmpty = FALSE, ...)
}
\arguments{
\item{Y}{an object describing the response variable. \code{Y} will be passed
to the \code{\link{plausible_predictor_test}} for analyzing.}

\item{X}{a matrix, data.frame or vector describing the covariates}

\item{E}{a vector, matrix or data frame describing the environments.}

\item{method}{a \strong{method object} greated by the
\code{\link{method_obj}} function. A method object is a list describing the
regression model, and the list has class equal to the name of the method
that should be invoked by the generic function
\code{\link{plausible_predictor_test}}. The \strong{method object} list
should always contain an entry named \code{model}, which equals a \code{R}
function (e.g. "glm", "coxph").}

\item{level}{numerical value between 0 and 1 denoting the significance level
used when testing. If not specified the algirithm will not return an
estimated set of identified causal predictors at level \code{level}.}

\item{fullAnalysis}{If \code{FALSE} those
  \code{\link{plausible_predictor_test}} that find p-values based on
  iterative test only test the hypothesis at the specified \code{level}.
  Hence \code{level} must be specified if \code{fullAnalysis} is set to
  \code{FALSE}.

  The inbuilt \code{\link{plausible_predictor_test}} for the
  \code{CRrectangle} and \code{CRellipsoid} method are both example of tests,
  that find the p-values using iterative tests. So if one of these methods is
  used, setting \code{fullAnalysis} to \code{FALSE} will save computational
  time. This does however also mean that it is not possible to estimate the
  p-values of the individual variables.}

\item{maxNoVariables}{The maximal number of variables in the tested subsets
of \code{X}. A smaller number saves computational time.}

\item{stopIfEmpty}{If \code{TRUE} the procedure will stop if the null
hypothesis for the empty set has been accepted. Setting to \code{TRUE} will
save computational time in these cases, but means that an analysis of other
subsets of \code{X} is lost.}

\item{...}{additional arguments carried to the
\code{\link{plausibel_predictor_test}}.}
}
\value{
Returns an object of class \code{ICP}. Such an object will at least
  contain the following

  \item{model.analysis }{a data frame listing the different models tested in
  the first column and the found p-values in the second column.}
  \item{call}{the matched call.}
  \item{level}{the significance level applied to all tests. If not specified
  this is simply \code{NULL}.}
  \item{method}{the method object used for the model fitting and hypothesis
  testing.}

  If a \code{level} has been specified then the following will also be part
  of the \code{ICP} object

  \item{empty.accepted  }{is 0 if the empty model was rejected and 1 if the
  empty model was accepted at level \code{level}}
  \item{accepted.model}{the estimated causal predictors}

  Lastly if \code{fullAnalysis} is set to \code{TRUE} then the \code{ICP}
  object will also contain

  \item{variable.analysis  }{a data frame with the predictor variables in the
  first column and their respective p-values found based on the model
  analysis in the second column.}
}
\description{
A method for finding causal predictors of a target variable.
}
\details{
The \code{ICP} function implements tooles for the \emph{Invariant Causal
Predictor} methodology described when the target variable is either described
by a glm, additive hazard model or multiplicative hazard model. As such the
target variable \code{Y} is allowed to be both a vector and a survival object.

The \code{ICP} function is essentially a wrapper function for the
\code{\link{plausible_predictor_test}} which tests the central null
hypothesis of the \emph{Invariance Causal Predictor} methodology, namely
\deqn{H_{0,S}: S is an invariant set w.r.t. (X,Y)}
where an invariant set is defined as a set of indicies \eqn{S} such that
\deqn{(Y^e | X^e_S) = (Y^f | X^f_S)}
in distribution for all environments \eqn{e,f}. If the data is time dependent
we formulate an analog invariance statement for all time points \eqn{t} and
\eqn{s}.

In the \code{ICPSurv} package 3 standard methods for the generic function
\code{\link{plausible_predictor_test}} has been implemented for testing the
null hypothesis above. For further discussion on how to implement new testing
methods for the above null hypothesis see
\code{\link{plausible_predictor_test}}.

To make the \code{ICP} function work correctly with the
\code{\link{plausible_predictor_test}} function a \strong{method object}
\code{method} must be specfied using the \code{\link{method_obj}} function.
The class of the \strong{method object} must correspond to an implemeted
method of the generic function \code{\link{plausible_predictor_test}}.
}
\examples{
# ===========================================================================
# A simple example with normal distributions and no time dependent variables
# First we simulate data.
G <- matrix(c(0,1,1,1,0,0,0,0,0,0.4,0,0,0,0,0.4,0,0,0,0,0,0,0,0,0,0),
            ncol = 5, byrow = T,
            dimnames = list(c("E","X1","X2","X3","Y"), c("E","X1","X2","X3","Y")))
sim <- c("sample(c(0,5,10),N,replace=T)", rep("rnorm(N,BETA,1)",4))
out <- list(Y = 5, X = 2:4, E = 1)
SIM <- sim_from_adj(G, 100, sim, out)
# note that Y has parents X1 and X2

# create list describing regression model
method <- list(model = "glm", family = "gaussian")

# analyze using the "Environment Irrelevance test"
class(method) <- "EnvirIrrel"
ICP(SIM$Y, SIM$X, SIM$E, method = method, level = 0.05)

# analyze using the old marginal confidence interval test
class(method) <- "CRrectangle"
ICP(SIM$Y, SIM$X, SIM$E, method = method, level = 0.05)

# analyze using the new "Confidence Region" test with correct ellipsoid CR's
class(method) <- "CRellipsoid"
ICP(SIM$Y, SIM$X, SIM$E, method = method, level = 0.05)
# ---------------------------------------------------------------------------



# ===========================================================================
# A simple example with poisson distributions and no time dependent variables
# First we simulate data.
G <- matrix(c(0,1,1,1,0,0,0,0,0,0.4,0,0,0,0,0.4,0,0,0,0,0,0,0,0,0,0),
            ncol = 5, byrow = T,
            dimnames = list(c("E","X1","X2","X3","Y"), c("E","X1","X2","X3","Y")))
sim <- c("sample(c(0,5,10),N,replace=T)", rep("rnorm(N,BETA,1)",3), "rpois(N,exp(BETA))")
out <- list(Y = 5, X = 2:4, E = 1)
SIM <- sim_from_adj(G, 100, sim, out)
# note that Y has parents X1 and X2

# create list describing regression model
method <- list(model = "glm", family = "poisson")

# analyze using the "Environment Irrelevance test"
class(method) <- "EnvirIrrel"
ICP(SIM$Y, SIM$X, SIM$E, method = method, level = 0.05)

# analyze using the old marginal confidence interval test
class(model) <- "CRrectangle"
ICP(SIM$Y, SIM$X, SIM$E, method = method, level = 0.05)

# analyze using the new "Confidence Region" test with correct ellipsoid CR's
class(method) <- "CRellipsoid"
ICP(SIM$Y, SIM$X, SIM$E, method = method, level = 0.05)
# ---------------------------------------------------------------------------



}
\references{
Peters, Jonas, Peter Bühlmann, and Nicolai Meinshausen. \emph{Causal
  inference by using invariant prediction: identification and confidence
  intervals.} Journal of the Royal Statistical Society: Series B (Statistical
  Methodology) 78.5 (2016): 947-1012.
}
\seealso{
\code{\link{print.ICP}} for summaries.
  \code{\link{plausible_predictor_test}} for customizing the \code{ICP}
  function.
}
