% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMR-main.R
\name{IMR}
\alias{IMR}
\title{Invariant Model Ranking}
\usage{
IMR(M, method = "BS", ...)
}
\arguments{
\item{M}{a list of models.}

\item{method}{bla bla bla}

\item{...}{TODO}
}
\value{
DESCRIPTION OF THE RETUNED
}
\description{
A method for ranking models based on invariance.
}
\details{
LONG DESCRIPTION
}
