% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICP-plausible_predictor_test.R
\name{intersect_CR}
\alias{intersect_CR}
\alias{intersect_CR.marginal}
\alias{intersect_CR.pairwise}
\alias{intersect_CR.QCLP}
\alias{intersect_marginal_rectangle}
\alias{intersect_ellipsoid_pair}
\alias{intersect_ellipsoid}
\title{Internal helper functions for CR plausible_predictor_test method}
\usage{
intersect_CR(method, models, alpha, ...)

\method{intersect_CR}{marginal}(method, models, alpha, fullAnalysis = FALSE, tol = 2e-10, ...)

\method{intersect_CR}{pairwise}(method, models, alpha, fullAnalysis = FALSE, tol = 2e-10, ...)

\method{intersect_CR}{QCLP}(method, models, alpha, fullAnalysis = TRUE, ...)

intersect_marginal_rectangle(models, alpha)

intersect_ellipsoid_pair(M1, M2, alpha)

intersect_ellipsoid(models)
}
\arguments{
\item{method}{TODO}

\item{models}{TODO}

\item{alpha}{TODO}

\item{...}{TODO}

\item{fullAnalysis}{TODO}

\item{tol}{TODO}

\item{M1, M2}{TODO}
}
\description{
Internal helper functions for CR plausible_predictor_test method
}
