% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICP-methodobj.R
\name{method_obj}
\alias{method_obj}
\alias{print.method_obj}
\title{Creating a Method Object for ICP}
\usage{
method_obj(model = "glm", method = "EnvirRel", ...)

\method{print}{method_obj}(x, exclude = c("id", "tol", "call", "dist"), ...)
}
\arguments{
\item{model}{the regression model used in \code{\link{ICP}}. The model class
specified here must be interpretable by the generic function
\code{\link{fit_model}} or \code{\link{fit_nonparam_model}}. If
\code{model} is not specified it will be set equal to \code{glm} and
\code{family = "gaussian"} will be added to the \strong{method object}.}

\item{method}{the testing metod used in \code{\link{ICP}}. The method
specified here must be interpretable by the generic function
\code{\link{plausible_predictor_test}}. If nothing is specified
\code{method} is set to \code{EnvirRel}.}

\item{...}{further arguments to be passed to the
  \code{\link{plausible_predictor_test}}, \code{\link{fit_model}} or
  \code{\link{fit_nonparam_model}} functions.

  If \code{method} is set to "\code{\link[=plausible_predictor_test]{CR}}"
  then a \code{solver} can be specified. The standard solver is
  \code{\link[=plausible_predictor_test]{QC}}. However, if \code{method} is
  "\code{\link[=plausible_predictor_test]{TimeVar}}" then \code{n.sim} can
  be set to any integer larger then 50.

  If \code{model} is \code{glm} then a \code{family} should be specified
  (see \code{\link[stats]{family}}). If a \code{family} is not specified it
  is set to "\code{gaussian}".}

\item{x}{a \code{method_obj} for printing.}

\item{exclude}{variables from the \code{method_obj} \code{x} which is not to
be included in the printout.}
}
\value{
The \code{method_obj} function returns a list with all the input
  parameters and their values. This list has class "method_obj",
  \code{method} and \code{model}. If \code{method} is \code{CR} then the
  output list also has class equal to \code{solver}. If, however, the
  \code{method} is \code{TimeVar} then the output also has class
  \code{nonparamtest}.
}
\description{
Creates a \strong{method object} within the \code{\link{ICP}} function.
}
\examples{
# The standard method object is a gaussian glm with method EnvirIrrel
ICPSurv:::method_obj()

# A method object for Intersecting Confidence Region analysis of linear models
ICPSurv:::method_obj(model = "lm", method = "CR", solver = "QC", fullAnalysis = TRUE)
# Here 'fullAnalysis' has been set to TRUE which will ensure that QCLP method
# does not use rectangle approximations as a first step

}
\seealso{
\code{\link{ICP}}, \code{\link{plausible_predictor_test}},
  \code{\link{fit_model}} or \code{\link{fit_nonparam_model}}  for use of the
  method object.
}
\keyword{internal}
