% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICP-variable_analysis.R
\name{model_analysis}
\alias{model_analysis}
\title{Find Accepted Model From ICP Object}
\usage{
model_analysis(x, level = 0.05, gof = 0.01)
}
\arguments{
\item{x}{An ICP object}

\item{level}{The significance level required for a model to be considered
invariant. The higher the significance \code{level} the stronger the
requirements for godness of fit to attain 'invariance' status.}

\item{gof}{If no set of variables (including the empty set) leads to a
p-value larger than the goodness-of-fit cutoff \code{gof}, the whole model
will be rejected. If the model is correct, this will happen with a
probability of gof and this option protects again making statements when
the model is obviously not suitable for the data.}
}
\value{
\code{model_analysis} returns an object of \code{\link[base]{class}}
"\code{model_analysis}" containing the following components:
\item{gof }{ goodness-of-fit cufoff.}
\item{level }{ significance level of hypothesis tests.}
\item{accepted.model }{ the accepted model.}
\item{empty.message }{ if the empty set is returned as \code{accepted.model}
  then \code{empty.message} will give detailes.}
\item{call }{ the matched call.}
}
\description{
The \code{model_analysis} function takes an \code{\link{ICP}} object and a
significance level and outputs the accepted model at this level.
}
\details{
The function \code{model_analysis} takes an \code{\link{ICP}} object and a
significance level and outputs the accepted model at this level:
\ifelse{html}{\out{<center>Accepted X columns = &#8745<sub>{S: H<sub>0,S</sub> accepted}</sub> S.</center>}}{\deqn{\textrm{Accepted } X \textrm{columns} = \cap_{\{S:\ H_{0,S} \textrm{ accepted}\}}\ S}}
See \code{\link{ICP}} for more details on the hypotheses
\ifelse{html}{\out{(H<sub>0,S</sub>)}}{\eqn{(H_{0,S})}}.

The \code{gof} parameter protects against making statements when the model is
obviously not suitable for the data. If no model reaches the threshold
\code{gof} significance level, i.e. the p-values for
\ifelse{html}{\out{(H<sub>0,S</sub>)}}{\eqn{(H_{0,S})}} are all smaller then
\code{gof}, we report that there is no evidence for an invariant set.

This function is also used internally in the \code{\link{ICP}} function
itself if \code{\link{ICP}} is called with \code{level} specified.
}
\examples{
n <- 100
E <- sample(5L, n, replace = TRUE)
X <- data.frame(X1 = rnorm(n, E, 1), X2 = rnorm(n, 3, 1))
Y <- rnorm(n, X$X1, 1)

obj <- ICP(Y, X, E, fullAnalysis = TRUE)
model_analysis(obj, level = 0.05, gof = 0.15)

}
\seealso{
The \code{model_analysis} function is also used internally in the
  \code{\link{ICP}} function itself if \code{\link{ICP}} is called with
  \code{level} specified.

  \code{variable_analysis} is another function for summarizing
  \code{\link{ICP}} objects.
}
