% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICP-plausible_predictor_test.R
\name{plausible_predictor_test}
\alias{plausible_predictor_test}
\alias{plausible_predictor_test.default}
\alias{plausible_predictor_test.nonparam}
\alias{plausible_predictor_test.EnvirIrrel}
\alias{plausible_predictor_test.CR}
\title{Test for plausible predictors}
\usage{
plausible_predictor_test(method, Y, X, ...)

\method{plausible_predictor_test}{default}(method, Y, X, ...)

\method{plausible_predictor_test}{nonparam}(method, Y, X, level, Bonferroni = TRUE, ...)

\method{plausible_predictor_test}{EnvirIrrel}(method, Y, X, E, ...)

\method{plausible_predictor_test}{CR}(
  method,
  Y,
  X,
  E,
  level,
  fullAnalysis,
  Bonferroni = TRUE,
  ...
)
}
\arguments{
\item{method}{a method object describing the model class to use in the
fitting procedure and the testing method. When used in the
\code{\link{ICP}} function the method object is inherited from the
\code{\link{ICP}} function.}

\item{Y}{a vector or \code{\link[survival]{Surv}} object describing the target variable. The \code{Y}
will be passed to the \code{\link{fit_model}} function for fitting, and it
is therefor important that the class of \code{Y} is understood by the
regression method used in \code{\link{fit_model}}. So if \code{method}
specifies that a cox regression is to be used, then \code{Y} must be a
\code{\link[survival]{Surv}} object.}

\item{X}{a matrix, vector or data frame describing the covariates. WRITE
ABOUT \code{X} WHEN TESTING THE EMPTY SET !!!!}

\item{...}{TODO}

\item{level}{the alpha level of testing, but it is only relevant if the
method is iterative \emph{and} \code{fullAnalysis} is \code{FALSE}.}

\item{Bonferroni}{TODO}

\item{E}{a vector describing the environmants}

\item{fullAnalysis}{TODO !!!!}
}
\value{
Returns a p-value, i.e. a number between 0 and 1, to be used in the
  \code{\link{ICP}} function.
}
\description{
\code{plausible_predictor_test} is a generic function used to test whether a
set of predictors might be plausible causal predictors.
}
\details{
The \code{plausible_predictor_test} is meant to be used in the wrapper
function \code{\link{ICP}}, but can also be used on its own as a simple
hypothesis testing function. The \strong{method object} \code{method}, which
is created by the \code{\link{method_obj}} function dictates the
\code{plausible_predictor_test} method. As this function is generic which
means that new testing methods can easily be added for new classes.

For the \code{plausible_predictor_test} to work correctly in the
\code{\link{ICP}} wrapper it must be a test of the null hypothesis
\deqn{H_{0,S}: S is an invariant set w.r.t. (X,Y)}
where an invariant set is defined as a set of indicies \eqn{S} such that
\deqn{(Y^e | X^e_S) = (Y^f | X^f_S)}
in distribution for all environments \eqn{e,f}. If the data is time dependent
we formulate an analog invariance statement for all time points \eqn{t} and
\eqn{s}. For more detailes and discussion of this hypothesis and invariance
see the references.

As the \code{plausible_predictor_test} is a hypothesis test it must return a
p value, that is a number between 0 and 1.

The \code{plausible_predictor_test}

PERHAPS A DESCRIPTION OF THE TACHNICALITIES OF \code{method} AND \code{Y} ?!
}
\seealso{
\code{\link{ICP}} for the full wrapper function.
}
