% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICP-variable_analysis.R
\name{variable_analysis}
\alias{variable_analysis}
\title{Determine Significance Level of Individual Variables}
\usage{
variable_analysis(x, gof = 0.1)
}
\arguments{
\item{x}{An ICP object}

\item{gof}{If no set of variables (including the empty set) leads to a
p-value larger than the goodness-of-fit cutoff \code{gof}, the whole model
will be rejected. If the model is correct, this will happen with a
probability of gof and this option protects again making statements when
the model is obviously not suitable for the data.}
}
\value{
\code{variable_analysis} returns a table with \code{\link[base]{class}}
  "\code{variable_analysis}" where each row reports the significance of the
  corresponding variable.
}
\description{
The \code{variable_analysis} function takes an \code{\link{ICP}} object and
outputs evidence for each individual variable being a direct cause.
}
\details{
The function \code{variable_analysis} takes an \code{\link{ICP}} object and a
minimum goodness-of-fit cutoff \code{gof} and outputs the evidence for each
individual variable being a direct cause. To conduct a variable analysis the
inputted \code{\link{ICP}} object should be the result of a call to the
\code{\link{ICP}} function with option \code{fullAnalysis = TRUE}. This
ensures that \code{variable_analysis} has access to the p-values for all the
hypothses \ifelse{html}{\out{(H<sub>0,S</sub>)}}{\eqn{(H_{0,S})}} (see
\code{\link{ICP}} for an explanation of the hypotheses).

If the needed \code{gof} level is reached the significance is calculated for
each variable \ifelse{html}{\out{X<sup>i</sup>}}{\eqn{X^i}} by
\ifelse{html}{\out{<center>p<sub>i</sub> = max{ p-value for H<sub>0,S</sub> &#8739 X<sup>i</sup> is not part of X<sup>S</sup> }.</center>}}{\deqn{p_i = \max\{\textrm{p-values for }H_{0,S}\ |\ X^i \textrm{ is not part of } X^S\}.}}

The \code{gof} parameter protects against making statements when the model is
obviously not suitable for the data. If no model reaches the threshold
\code{gof} significance level, i.e. the p-values for
\ifelse{html}{\out{(H<sub>0,S</sub>)}}{\eqn{(H_{0,S})}} are all smaller then
\code{gof}, we report that there is no evidence for individual variables, as
there is no evidence for an invariant set.

The \code{variable_analysis} function is also used internally in the
\code{\link{ICP}} function itself if \code{\link{ICP}} is called with
\code{level} specified and \code{fullAnalysis = TRUE}.
}
\examples{
n <- 100
E <- sample(5L, n, replace = TRUE)
X <- data.frame(X1 = rnorm(n, E, 1), X2 = rnorm(n, 3, 1))
Y <- rnorm(n, X$X1, 1)

obj <- ICP(Y, X, E, level = 0.05, fullAnalysis = TRUE)
variable_analysis(obj, gof = 0.1)

# If we make extreem requirements for gof
# all "p-values" will be taken to be 1,
# i.e. no variable is a significant causal predictor
variable_analysis(obj, gof = 0.9)

}
\seealso{
The \code{variable_analysis} function is also used internally in the
  \code{\link{ICP}} function itself if \code{\link{ICP}} is called with
  \code{fullAnalysis = TRUE}.

  \code{model_analysis} is another function for summarizing
  \code{\link{ICP}} objects.
}
