% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMR-analysis_constructors.R
\name{construct_analysis}
\alias{construct_analysis}
\alias{construct_analysis.default}
\alias{construct_analysis.BS}
\title{Internal Methods for Constructing IMR Analyses}
\usage{
\method{construct_analysis}{default}(mc, data, ...)

\method{construct_analysis}{BS}(mc, data, norder, knots, L, ...)
}
\arguments{
\item{mc}{the classes of the fitted models to be analyzed via the constructed
analysis funciton.}

\item{data}{data from the first fitted model.}

\item{...}{additional arguments.}
}
\value{
An analysis constructor must return the following in a list
\item{analysis }{ a function with exactly one input, \code{data}, which has the
same format as the inputted \code{data}.}
\item{method }{ a method list with info used in the \code{\link{plot.IMR}}
and \code{print.IMR} functions.}
}
\description{
An internal method used in the \code{\link{IMR}} for construction of the
model analysis.
}
\details{
\code{construct_analysis} is a generic function. The methods should
correspond to the inputted \code{method} in \code{\link{IMR}}.
}
\keyword{internal}
