% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMR-main.R
\name{construct_analysis}
\alias{construct_analysis}
\alias{construct_analysis.default}
\alias{construct_analysis.BS}
\alias{construct_analysis.KS}
\alias{construct_analysis.CM}
\title{Internal Method for Constructing IMR Analysis}
\usage{
construct_analysis(mc, data, ...)

\method{construct_analysis}{default}(mc, data, ...)

\method{construct_analysis}{BS}(mc, data, norder, knots, L, W, ...)

\method{construct_analysis}{KS}(mc, X, W, ...)

\method{construct_analysis}{CM}(mc, X, ...)
}
\arguments{
\item{mc}{TODO}

\item{data}{TODO}

\item{...}{TODO}

\item{norder}{TODO}

\item{knots}{TODO}

\item{L}{TODO}

\item{W}{TODO}
}
\value{
a function that has input \code{Y}.
}
\description{
An internal method used in the \code{\link{IMR}} for construction of the
model analysis.
}
\details{
LONG DESCRIPTION
}
