% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMR-analysis_constructors.R
\name{construct_analysis}
\alias{construct_analysis}
\alias{construct_analysis.default}
\alias{construct_analysis.BS}
\title{Internal Method for Constructing IMR Analysis}
\usage{
construct_analysis(mc, data, ...)

\method{construct_analysis}{default}(mc, data, ...)

\method{construct_analysis}{BS}(mc, data, norder, knots, L, ...)
}
\arguments{
\item{mc}{TODO}

\item{data}{TODO}

\item{...}{TODO}

\item{norder}{TODO}

\item{knots}{TODO}

\item{L}{TODO}
}
\value{
a function that has input \code{Y}.
}
\description{
An internal method used in the \code{\link{IMR}} for construction of the
model analysis.
}
\details{
LONG DESCRIPTION
}
\seealso{
\code{\link{IMR}}
}
