% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICP-fitters.R
\name{fit_model}
\alias{fit_model}
\alias{fit_model.default}
\alias{fit_model.lm}
\alias{fit_model.glm}
\alias{fit_model.ph}
\alias{fit_model.ah}
\alias{fit_model.hazard}
\alias{fit_nonparam_model}
\alias{fit_nonparam_model.default}
\alias{fit_nonparam_model.ph}
\alias{fit_nonparam_model.ah}
\alias{fit_nonparam_model.hazard}
\title{Internal regression functions}
\usage{
fit_model(method, Y, X, ...)

\method{fit_model}{default}(method, Y, X, ...)

\method{fit_model}{lm}(method, Y, X, ...)

\method{fit_model}{glm}(method, Y, X, ...)

\method{fit_model}{ph}(method, Y, X, ...)

\method{fit_model}{ah}(method, Y, X, ...)

\method{fit_model}{hazard}(method, Y, X, ...)

fit_nonparam_model(method, Y, X, ...)

\method{fit_nonparam_model}{default}(method, Y, X, ...)

\method{fit_nonparam_model}{ph}(method, Y, X, id = NULL, ...)

\method{fit_nonparam_model}{ah}(method, Y, X, id = NULL, ...)

\method{fit_nonparam_model}{hazard}(method, Y, X, id = NULL, ...)
}
\arguments{
\item{method}{a \strong{method object} created by the
\code{\link{method_obj}} function.}

\item{Y}{a vector or \code{\link[survival]{Surv}} object describing the
target variable.}

\item{X}{a matrix, vector or data frame describing the covariates. WRITE
ABOUT \code{X} WHEN TESTING THE EMPTY SET !!!!}

\item{...}{additional arguments. At the moment only the additional argument
\code{const} is allowed.}

\item{id}{TODO}
}
\value{
Depending on \code{method} the \code{fit_model} function returns
  different lists. If \code{method$method} \eqn{!=} "\code{ConstTime}" the output list must
  contain the following
  \describe{
    \item{coefficients}{the estimated regression coefficients.}
    \item{covariance}{the estimated covariance matrix.}
    \item{deviance}{the deviance, i.e. \eqn{- 2 * LogLik}, where \eqn{LogLik}
      is the log likelihood evaluated in the reported estimators.}
    \item{df}{degrees of freedom}
  }

  If \code{method} has class "\code{ConstTime}" then the output list contains
  ANNA: DO WE ALWAYS RETURN ALL 4 ??? If we don't need pvals then we may set
  n.sim = 0

  \describe{
    \item{cum}{cumulative timevarying regression coefficients estimated within
      the timeinterval with observations.}
    \item{var.cum}{rhe martingale based pointwise variance estimated for
      cumulatives.}
    \item{pval.testBeqC}{Kolmogorov-Smirnov test p-values based on resampling.}
    \item{pval.testBeqC.is}{Carmer von Mises test p-values based on resampling.}
   }
}
\description{
The \code{fit_model} function is a generic function meant for internal use in
the \code{ICPSurv} package, and as such it is not exported to the namespace.
}
\details{
MORE DESCRIPTION

MENTION SPECIAL CASE COX AND AALEN NOT CONST.
}
\seealso{
\code{\link{ICP}}, \code{\link{plausible_predictor_test}} for the
  full wrapper functions.
}
