% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICP-main.R
\name{ICP}
\alias{ICP}
\title{Invariant Causal Prediction}
\usage{
ICP(
  Y,
  X,
  E = NULL,
  model = "lm",
  method = "EnvirIrrel",
  level = 0.05,
  maxNoVariables = 8,
  fullAnalysis = TRUE,
  ...
)
}
\arguments{
\item{Y}{an object describing the response variable. \code{Y} will be passed
to the \code{\link{plausible_predictor_test}} for analyzing.}

\item{X}{a matrix, data.frame or vector describing the covariates}

\item{E}{a vector, matrix or data frame describing the environments.}

\item{method}{a character which must correspond to a
\code{\link{plausible_predictor_test}} method. In this package the methods
\code{\link[=plausible_predictor_test.EnvirIrrel]{EnvirIrrel}},
\code{\link[=plausible_predictor_test.CR]{CR}} and
\code{\link[=plausible_predictor_test.nonparam]{nonparam}} are already
defined. If the user defines a new S3 method for the
\code{\link{plausible_predictor_test}} then this new method becomes a valid
input for \code{method} here. For more details on how to implement your own
method for use in the \code{ICP} function see
\code{\link{plausible_predictor_test}}.}

\item{level}{numerical value between 0 and 1 denoting the significance level
used when testing. If not specified the algirithm will only calculate the
p-values of the null hypothesis \eqn{H_{0,S}} and draw no conclusions based
on these values.}

\item{maxNoVariables}{The maximal number of variables in the tested subsets
of \code{X}. A smaller number saves computational time, but also
corresponds to assuming that the target variable \code{Y} has at most
\code{maxNoVariables} identifiable predictors.}

\item{fullAnalysis}{if \code{FALSE} different computations time saving
options will be enabled and as a result many methods will not be able to
return proper p-values of the null hypothesis \eqn{H_{0,S}}. In turn this
means that the method will no longer be able to estime the significanse of
a single variable. \code{fullAnalysis = FALSE} will result in different
behavior for different \code{methods}, so for a more detailed discussion of
concrete effects see the help page for the chooosen \code{method}.}

\item{...}{additional arguments carried to the
\code{\link{plausible_predictor_test}}.}
}
\value{
Returns an object of class \code{ICP}. Such an object will at least
  contain the following

  \item{model.analysis }{a data frame listing the different models tested in
  the first column and the found p-values in the second column.}
  \item{call}{the matched call.}
  \item{level}{the significance level applied to all tests. If not specified
  this is simply \code{NULL}.}
  \item{method}{the method object used for the model fitting and hypothesis
  testing.}

  If a \code{level} has been specified then the following will also be part
  of the \code{ICP} object

  \item{empty.accepted  }{is 0 if the empty model was rejected and 1 if the
  empty model was accepted at level \code{level}}
  \item{accepted.model}{the estimated causal predictors}

  Lastly if \code{fullAnalysis} is set to \code{TRUE} then the \code{ICP}
  object will also contain

  \item{variable.analysis  }{a data frame with the predictor variables in the
  first column and their respective p-values found based on the model
  analysis in the second column.}
}
\description{
A method for finding causal predictors of a target variable.
}
\details{
The \code{ICP} function implements tooles for the \emph{Invariant Causal
Predictor} methodology when the target variable may be described by a lm,
glm or hazard model. To allow for survival type models the target variable
\code{Y} is allowed to be both a vector and a survival object.

The \code{ICP} function is essentially a wrapper function for the
\code{\link{plausible_predictor_test}} which tests the central null
hypothesis of the \emph{Invariance Causal Predictor} methodology, namely
\deqn{H_{0,S}: S is an invariant set w.r.t. (X,Y)}
where an invariant set is defined as a set of indicies \eqn{S} such that
\deqn{(Y^e | X^e_S) = (Y^f | X^f_S)}
in distribution for all environments \eqn{e,f}. If the data is time dependent
we formulate an analog invariance statement for all time points \eqn{t} and
\eqn{s}.

In the \code{ICPSurv} package three standard methods for the generic function
\code{\link{plausible_predictor_test}} has been implemented for testing the
null hypothesis above. These are the
\code{\link[=plausible_predictor_test]{EnvirIrrel}},
\code{\link[=plausible_predictor_test]{CR}} and
\code{\link[=plausible_predictor_test]{nonparam}} method. For further
discussion on how to implement new testing methods for the above null
hypothesis see \code{\link{plausible_predictor_test}}.
}
\examples{
# ===========================================================================
# An example with normal distributions
n <- 500
E <- sample(5L, n, replace = TRUE)
X <- rnorm(n, E, 1)
Y <- rnorm(n, X, 1)
ICP(Y, X, E) # Environment Irrelevance Test
ICP(Y, X, E, method = "CR") # Confidence Region Test

# ===========================================================================
# An example with a poisson distribution
Y <- rpois(n, exp(X))
# Environment Irrelevance Test
ICP(Y, X, E, model = "glm", family = "poisson")
# Intersecting Confidence Region Test
ICP(Y, X, E, model = "glm", family = "poisson", method = "CR")

# ===========================================================================
# An example with right censored survival times
Y <- rexp(n, exp(- 1.5 *X))
C <- rexp(n, exp(- 2.0 * X))
time <- pmin(Y, C)
status <- time == Y
# Environment Irrelevance Test
ICP(survival::Surv(time, status), X, E, model = "ph")
# Intersecting Confidence Regions Test
ICP(survival::Surv(time, status), X, E, model = "ph", method = "CR")
# Non-parametric Constant Effect Over Time Test
ICP(survival::Surv(time, status), X, E, model = "ph", method = "nonparam")

}
\references{
Jonas Peters, Peter Bühlmann, and Nicolai Meinshausen. \emph{Causal
  inference by using invariant prediction: identification and confidence
  intervals.} Journal of the Royal Statistical Society: Series B (Statistical
  Methodology) 78.5 (2016): 947-1012.
}
\seealso{
\code{\link{print.ICP}} for summaries.
  \code{\link{plausible_predictor_test}} for customizing the \code{ICP}
  function.
}
