% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICP-main.R
\name{method_obj}
\alias{method_obj}
\title{Creating a Method Object for ICP}
\usage{
method_obj(model = "glm", method = "EnvirIrrel", ...)
}
\arguments{
\item{model}{the regression model used in \code{\link{ICP}}. The model class
specified here must be interpretable by the generic function
\code{\link{fit_model}}. If \code{model} is not specified it will be set
equal to \code{glm} and \code{family = "gaussian"} will be added to the
\strong{method object}.}

\item{method}{the testing metod used in \code{\link{ICP}}. The method
specified here must be interpretable by the generic function
\code{\link{plausible_predictor_test}}. If nothing is specified
\code{method} is set to \code{EnvirIrrel}}

\item{...}{further arguments to be passed to the
\code{\link{plausible_predictor_test}} or \code{\link{fit_model}}
functions.}
}
\value{
a list with all the input parameters and their values. This list has
  class "Method_obj", \code{method} and \code{model}.
}
\description{
THIS FUNCTION IS EXPORTED - IS THIS A GOOD IDEA WITH THIS NAME ?? Creates a
\strong{method object} compatible with the \code{\link{ICP}} function.
}
\examples{
# the standard method object is of method "EnvirIrrel"
method_obj()

# create a method object suitable for "confidence region" testing
# with a poisson glm regression model
method_obj(method = "CRellipsoid", model = "glm", family = "poisson")

# create a method object suitable for "environment irrelevance" testing
# with a cox regression model
method_obj(method = "EnvirIrrel", model = "survival::coxph")
# if the survival package is attached the 'survival::' part can be dropped

}
\seealso{
\code{\link{ICP}}, \code{\link{plausible_predictor_test}},
  \code{\link{fit_model}} for use of the method object.
}
