% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMR-main.R
\name{plot.IMR}
\alias{plot.IMR}
\title{Plot IMR Analysis}
\usage{
\method{plot}{IMR}(x, y = NULL, which = seq_along(x$results), ...)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{which}{TODO}

\item{...}{TODO}
}
\value{
DESCRIPTION OF RETURNED
}
\description{
THIS FUNCTION IS DEFINITELY NOT DONE !!
}
\details{
LONG DESCRIPTION OF PLOTTER FUNCTION
}
\examples{
# Make super good example

}
