% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICP-summary_function.R
\name{print.ICP}
\alias{print.ICP}
\title{Printing an ICP output}
\usage{
\method{print}{ICP}(x, ...)
}
\arguments{
\item{x}{an object of class "ICP", usually, a result of a call to
\code{\link{ICP}}}

\item{...}{TODO}
}
\description{
A simple print function for the \code{\link{ICP}} function.
}
\seealso{
The Invariant Causal Prediction function \code{\link{ICP}}.
}
